+++ src/Network/Network/Models/NetworkManager/PSNetworkManagerScopeConnection.cs
@@ -27,7 +27,5 @@ namespace Microsoft.Azure.Commands.Network.Models.NetworkManager
         public string ResourceId { get; set; }
 
         public string ConnectionState { get; set; }
-
-        public new string Description { get; set; }
     }
 }